Added abstract notification class: new notification interfaces to be implemented.
