+++ config.cfg
@@ -1,3 +1,3 @@
 [main]
-repository_path = /home/shuma/erp_modules/kt/,/home/shuma/erp_modules/wr/,/home/shuma/erp_modules/pinnacle/,/home/shuma/work/ea_web,/home/shuma/work/openerp/
+repository_path = /Users/ngud/git_notifier/
 repeat_every = 20
